# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: PVA Bazaar Marketplace API
  description: Complete global marketplace for artisans, creators, and sellers with all 8 phases
  version: 8.0.0
  contact:
    name: PVA Support
    url: https://pvabazaar.org
  license:
    name: Proprietary
    url: https://pvabazaar.org/license
servers:
  - url: https://api.pvabazaar.org
    description: Production
  - url: http://localhost:5001
    description: Development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API Key for partner integrations
...
security:
  - bearerAuth: []
paths:
...
  /api/fulfillment/select-center:
    post:
      tags: [Fulfillment - Phase 6b]
      summary: Select best fulfillment center
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                destinationCountry: { type: string }
                destinationCity: { type: string }
      responses:
        '200':
          description: Selected fulfillment center
...
  /api/fulfillment/create-shipment:
    post:
      tags: [Fulfillment - Phase 6b]
      summary: Create shipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Shipment created
  /api/fulfillment/track-shipment/{trackingNumber}:
    get:
      tags: [Fulfillment - Phase 6b]
      summary: Track shipment
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Shipment tracking info
  /api/fulfillment/initiate-return:
    post:
      tags: [Fulfillment - Phase 6b]
      summary: Initiate return process
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId: { type: string }
                reason: { type: string }
      responses:
        '200':
          description: Return initiated

  # ===== PHASE 7: AI HELPER =====
  /api/ai-help/ask:
    post:
      tags: [AI Help - Phase 7]
      summary: Ask AI assistant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question: { type: string }
                topic: { type: string }
      responses:
        '200':
          description: AI response
  /api/ai-help/guides/{topic}:
    get:
      tags: [AI Help - Phase 7]
      summary: Get help guide for topic
      parameters:
        - name: topic
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Guide content
  /api/ai-help/pricing-suggest:
    post:
      tags: [AI Help - Phase 7]
      summary: Get AI pricing suggestions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Pricing suggestions
  /api/ai-help/compliance-checklist:
    get:
      tags: [AI Help - Phase 7]
      summary: Get compliance checklist
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Compliance items

  # ===== PHASE 8: OPEN API =====
...