# backend/routes/ai-help.js
// backend/routes/ai-help.js - AI-powered assistance for sellers and buyers
const express = require('express');
const router = express.Router();
const { authenticateToken } = require('../middleware/auth');
const AIAgentBot = require('../models/AIAgentBot');
const Shop = require('../models/Shop');
const SellerAnalytics = require('../models/SellerAnalytics');
const ProductType = require('../models/ProductType');
const pricingService = require('../services/pricingService');

// Placeholder for Claude API integration (replace with actual API call)
const generateAIResponse = async (messages, topic) => {
  // In production, integrate with Anthropic Claude API here
  const prompts = {
    pricing: `You are a marketplace pricing expert. Help sellers set competitive prices based on market data, costs, and demand.`,
    registration: `You are a helpful guide for new marketplace sellers. Help them understand account setup, verification, and shop creation.`,
    shipping: `You are a logistics expert. Help sellers understand shipping options, rates, fulfillment centers, and international delivery.`,
    compliance: `You are a legal compliance advisor. Help sellers understand tax obligations, business registration, and platform policies.`,
    marketing: `You are a marketplace marketing specialist. Help sellers improve sales through better product descriptions, promotions, and visibility.`,
    onboarding: `You are an onboarding specialist. Guide new users step-by-step through marketplace registration and first steps.`,
  };

  // Mock response (replace with actual Claude API)
  return {
    message: `I'd be happy to help with ${topic}. Based on the information provided, here are my recommendations: [AI response would be generated by Claude API]`,
    suggestions: ['Suggestion 1', 'Suggestion 2', 'Suggestion 3'],
  };
};

/**
 * Ask AI assistant a question
 * POST /api/ai-help/ask
 */
router.post('/ask', authenticateToken, async (req, res) => {
  try {
    const { question, topic = 'general', sessionId } = req.body;
    const userId = req.user.id;

    if (!question) {
      return res.status(400).json({ error: 'question required' });
    }

    // Find or create session
    let session = sessionId ? await AIAgentBot.findById(sessionId) : null;

    if (!session) {
      session = new AIAgentBot({
        userId,
        topic,
        sessionId: `${userId}-${Date.now()}`,
      });
    }

    // Add user message
    session.conversation.push({
      role: 'user',
      message: question,
    });

    // Generate AI response
    const aiResponse = await generateAIResponse(session.conversation, topic);

    session.conversation.push({
      role: 'assistant',
      message: aiResponse.message,
    });

    await session.save();

    res.json({
      sessionId: session._id,
      response: aiResponse.message,
      suggestions: aiResponse.suggestions,
    });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

/**
 * Get help guide for a topic
 * GET /api/ai-help/guides/:topic
 */
router.get('/guides/:topic', async (req, res) => {
  try {
    const { topic } = req.params;

    const guides = {
      pricing: {
        title: 'Pricing Guide',
        sections: [
          {
            heading: 'Understanding Market Rates',
            content:
              'Research similar products to understand the competitive landscape. Consider production costs, desired profit margin, and market demand.',
          },
          {
            heading: 'Dynamic Pricing',
            content:
              'Adjust prices seasonally and based on demand. Use our fair price calculator to get recommendations based on real market data.',
          },
          {
            heading: 'Avoid Common Mistakes',
            content:
              'Do not underprice dramatically unless running a promotion. Do not price based on wishful thinking. Use data-driven decisions.',
          },
        ],
        tools: [
          'Fair Price Calculator',
          'Market Intelligence Dashboard',
          'Competitor Price Tracker',
        ],
      },
      registration: {
        title: 'Registration & Account Setup',
        sections: [
          {
            heading: 'Create Your Account',
            content:
              'Sign up with your email, verify ownership, and complete KYC verification for seller privileges.',
          },
          {
            heading: 'Setting Up Your Shop',
            content:
              'Create a shop storefront with branding (logo, banner, colors), shop description, and policies (shipping, returns).',
          },
          {
            heading: 'Payment Methods',
            content:
              'Connect a bank account for payouts. Choose payment processors (Stripe, crypto, local methods). Set up payment splits if applicable.',
          },
        ],
        tools: ['Shop Builder Wizard', 'Bank Connection Tool', 'Verification Center'],
      },
      shipping: {
        title: 'Shipping & Fulfillment',
        sections: [
          {
            heading: 'Choose a Fulfillment Center',
            content:
              'Select fulfillment centers by location. We recommend the closest center to your customers for faster delivery.',
          },
          {
            heading: 'Shipping Methods',
            content:
              'Offer multiple shipping options (standard, express, overnight). Get rate quotes from multiple carriers.',
          },
          {
            heading: 'International Shipping',
            content:
              'For global sales, use our international fulfillment network. We handle customs, duties, and local delivery partners.',
          },
        ],
        tools: ['Fulfillment Center Locator', 'Shipping Rate Calculator', 'Return Label Generator'],
      },
      compliance: {
        title: 'Legal & Tax Compliance',
        sections: [
          {
            heading: 'Business Registration',
            content:
              'Register your business appropriately for your jurisdiction (sole proprietor, LLC, corporation, cooperative, etc.).',
          },
          {
            heading: 'Tax Obligations',
            content:
              'Understand sales tax, VAT, and income tax requirements. We provide tax reports to help with filing. Consult a tax professional.',
          },
          {
            heading: 'Intellectual Property',
            content:
              'Ensure products are original or properly licensed. Register trademarks/copyrights if applicable. Respect platform IP policies.',
          },
        ],
        tools: ['Compliance Checklist', 'Tax Report Generator', 'Legal Template Library'],
      },
    };

    const guide = guides[topic] || guides['general'];
    res.json(guide);
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

/**
 * Get AI pricing suggestions
 * POST /api/ai-help/pricing-suggest
 */
router.post('/pricing-suggest', authenticateToken, async (req, res) => {
  try {
    const { productId, cost, targetMargin, condition, category } = req.body;
    const userId = req.user.id;

    if (!productId || !cost || !targetMargin) {
      return res.status(400).json({ error: 'productId, cost, targetMargin required' });
    }

    // Get fair price recommendation from pricing service
    const fairPrice = await pricingService.calculateFairPrice(productId, { condition, category });

    // Calculate target price with margin
    const targetPrice = Math.round(cost / (1 - targetMargin / 100));

    // Create recommendation
    const recommendation = {
      fairPrice: fairPrice.price,
      fairPriceRationale: `Market average: $${fairPrice.marketAvg}, range: $${fairPrice.minPrice} - $${fairPrice.maxPrice}`,
      recommendedPrice: Math.round((fairPrice.price + targetPrice) / 2), // Blend fair price with target
      targetPrice,
      urgencyOptions: {
        standard: fairPrice.price,
        promotional: Math.round(fairPrice.price * 0.85),
        premium: Math.round(fairPrice.price * 1.2),
        clearance: Math.round(fairPrice.price * 0.7),
      },
      estimatedSalesVelocity: {
        atFairPrice: '15-25 units/month (estimate)',
        atPromotionalPrice: '30-40 units/month',
        atPremiumPrice: '5-10 units/month',
      },
    };

    res.json(recommendation);
  } catch (error) {
    res.status(400).json({ error: error.message });
  }
});

/**
 * Get seller compliance checklist
 * GET /api/ai-help/compliance-checklist
 */
router.get('/compliance-checklist', authenticateToken, async (req, res) => {
  try {
    const userId = req.user.id;

    // Get seller profile for jurisdiction
    const shop = await Shop.findOne({ userId }).populate('userId', 'country');

    const checklist = [
      {
        title: 'Business Registration',
        status: 'pending',
        description: 'Register business with local authorities',
        deadline: '30 days',
        link: '/help/business-registration',
      },
      {
        title: 'Tax ID / EIN',
        status: 'pending',
        description: 'Obtain tax identification number',
        deadline: '60 days',
        link: '/help/tax-setup',
      },
      {
        title: 'Bank Account Connection',
        status: 'pending',
        description: 'Link verified bank account for payouts',
        deadline: '15 days',
        link: '/seller/payment-settings',
      },
      {
        title: 'Seller Agreement',
        status: 'pending',
        description: 'Accept marketplace seller terms and policies',
        deadline: '5 days',
        link: '/legal/seller-agreement',
      },
      {
        title: 'Product Compliance',
        status: 'pending',
        description: 'Ensure products comply with regulations (CE, FCC, etc.)',
        deadline: 'ongoing',
        link: '/help/product-compliance',
      },
      {
        title: 'Shipping Policy',
        status: 'pending',
        description: 'Define and publish shipping policies',
        deadline: '10 days',
        link: '/seller/settings/policies',
      },
      {
        title: 'Returns Policy',
        status: 'pending',
        description: 'Define customer returns and refund policy',
        deadline: '10 days',
        link: '/seller/settings/policies',
      },
      {
        title: 'Privacy Policy',
        status: 'pending',
        description: 'Create privacy policy for customer data handling',
        deadline: '15 days',
        link: '/help/privacy-policy-template',
      },
    ];

    res.json({
      completionPercentage: 0,
      checklist,
      nextSteps: ['Complete Business Registration', 'Obtain Tax ID', 'Connect Bank Account'],
    });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

/**
 * Get AI insights about seller performance
 * GET /api/ai-help/performance-insights
 */
router.get('/performance-insights', authenticateToken, async (req, res) => {
  try {
    const userId = req.user.id;

    // Get latest analytics
    const analytics = await SellerAnalytics.findOne({ sellerId: userId }).sort({ createdAt: -1 });

    if (!analytics) {
      return res.json({
        insights: ['Start by creating your first product to get performance insights'],
      });
    }

    const insights = [];

    // Conversion rate analysis
    if (analytics.conversionRate && analytics.conversionRate < 2) {
      insights.push({
        category: 'Conversion',
        priority: 'high',
        insight: `Your conversion rate is ${analytics.conversionRate}%, below market average 2-3%. Try: better product photos, detailed descriptions, customer testimonials.`,
      });
    }

    // Traffic analysis
    if (analytics.pageViews > 0 && analytics.checkoutStarts / analytics.pageViews < 0.1) {
      insights.push({
        category: 'Traffic',
        priority: 'medium',
        insight: `Only ${((analytics.checkoutStarts / analytics.pageViews) * 100).toFixed(1)}% of visitors view checkout. Improve: product descriptions, add reviews, reduce checkout friction.`,
      });
    }

    // Revenue growth
    if (analytics.totalRevenue === 0) {
      insights.push({
        category: 'Revenue',
        priority: 'high',
        insight: `No sales yet. Focus on: creating quality product listings, using relevant tags, and responding quickly to customer inquiries.`,
      });
    }

    // Pricing insights
    insights.push({
      category: 'Pricing',
      priority: 'medium',
      insight: `Use our Fair Price Calculator to ensure competitive pricing. Current avg order value: $${analytics.avgOrderValue || 0}`,
    });

    // Recommendations
    const recommendations = [
      'Add at least 10 high-quality product photos',
      'Write detailed product descriptions (200+ characters)',
      'Respond to inquiries within 1 hour',
      'Offer competitive shipping rates',
      'Ask satisfied customers for reviews',
    ];

    res.json({
      insights,
      recommendations,
      nextSteps: insights.filter((i) => i.priority === 'high').map((i) => i.insight),
    });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

/**
 * Rate AI assistance helpfulness
 * POST /api/ai-help/rate-session
 */
router.post('/rate-session', authenticateToken, async (req, res) => {
  try {
    const { sessionId, rating, comment } = req.body;

    if (!sessionId || !rating) {
      return res.status(400).json({ error: 'sessionId and rating required' });
    }

    if (rating < 1 || rating > 5) {
      return res.status(400).json({ error: 'rating must be 1-5' });
    }

    const session = await AIAgentBot.findByIdAndUpdate(
      sessionId,
      {
        resolved: true,
        resolvedAt: new Date(),
        satisfactionRating: rating,
        satisfactionComment: comment,
      },
      { new: true },
    );

    res.json({ success: true, session });
  } catch (error) {
    res.status(500).json({ error: error.message });
  }
});

module.exports = router;